JBIDE-21866 - liberalize some code and class structures to allow openshift to reuse
